feat: feito melhorias iniciais com o hr-payroll

+++ hr-payroll/src/main/java/com/devsuperior/hrpayroll/clients/worker/WorkerClient.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 @Component
 @FeignClient(name = "hr-worker", url = "localhost:8001", path = "/workers")
 public interface WorkerClient {
-    
+
     @GetMapping("/{id}")
-    ResponseEntity<Worker> getWorker(@PathVariable Integer id);
+    ResponseEntity<Worker> getWorker(@PathVariable Long id);
 }
